requirements.txt
snowflake-snowpark-python
pandas
polars
plotly
matplotlib
seaborn
SQLalchemy
ipykernel
scikit-learn